Add support for the user to control the grid spacing via a pop-up.
This goes together with changes to the class saavtk.model.Graticule. The
changes were all made together on the sbmt2 branch originally, but on
that branch there are actually TWO GraticulePopupMenu classes, and the
other one (saavtk.gui.popup.GraticulePopupMenu, in a different package)
was the one that was in use and had these changes.

So in porting this change back from sbmt2 into sbmt1dev, was able to
cherry-pick the changes to the Graticule class, but had to re-apply the
changes in parallel to *this* version of GraticulePopupMenu.