Import 2-stage constructor idea from saavtk2 as part of saavtk1dev.
Old constructors remain and are unaffected so sbmt1 patterns continue to
work. The new pattern allows for subclasses to call the superclass
constructor and then separately call the (new)
initializeConfigParameters method to do the rest of what the original
constructor does. This is to support changes to the ViewConfig family to
simplify and modularize how models are created.