Initial implementation of RenderPanel and RenderToolbar.
The coordinate axes are presently rendered offscreen and then mapped to
a quad that follows the camera; this suffers from minor distortion
effects due to perspective projection. The axes widget that vtk provides
is presently broken for vtkJoglPanelComponents, but we used to use
vtkJoglCanvasComponents which are heavyweight awt components and break
the jslidepane behavior.