Improved Renderer/RenderPanel and associated class.
- Redesigned the RenderPanel as a more 'primitive' component. It no
longer listens to the RenderToolbar or interacts with it directly. It no
longer responds to SAAVTK camera events.
- Redesigned the Renderer as a more 'primitive' component. It no longer
keeps a reference to the StatusBar. It holds a SAAVTK camera and
delegates some camera based logic to the camera.
- Eliminated some camera management code (moved/merged to camera
package).
- Reorganized some method declarations to be at more logical location in
source.
- Eliminated redundant / noop code / duplicate field members.
- Cleaned up / organized field declaration area.
- Updated logic so that axis rotation may be constrained to an arbitrary
axis rather than a standard coordinate axis. CartesianAxis /
CartesianViewDirection are no longer necessary as a result.