RM 1262: Detect when APL's server returns an HTML error file.
Sometimes, for reasons unkonwn, APL's server will not reject
a request by setting a proper URL code like 302. Instead,
it actually returns a file that contains HTML to display a
message to the effect of:

<html><head><title>Request Rejected</title></head><body>...

This code change checks for a string like this in the first
few bytes of the file, and it if finds it, it throws an
exception, thus preventing the HTML file from being downloaded
as if it were the real file being requested. This prevents the
previous (undesired) behavior in which the cache or custom
data area must be manually repaired to remove such files after
they were unintentionally downloaded.