RM 1422: several changes (see full commit message).
1. Added a field for enabling/disabling the server-side check.
This is just for internal debugging for now.
2. Added a forceUpdate parameter to the doAccessCheckOnServer
methods. This goes with #5.
3. Restore code in doAccessCheckOnServer to remove the
data root directory from URLs. This makes the queries smaller,
which will allow more URLs to be checked at one go.
4. Check for "Request Rejected" HTML in the stream returned
from the query, and throw an appropriate exception so this
doesn't get parsed as if it were valid input.
5. Update the file state as well as the URL state in
doAccessCheckOnServer. Otherwise the file state never gets
correctly updated, and things stay greyed out that should be
active.
6. Restore some of the Debug messages so that developers may
be aware when the server-side check is failing.
7. Apply name change FileAccessQuerier ->
DownloadableFileAccessQuerier.
8. Small clean-up items.