RM 2060: NOT READY FOR MERGE -- SEE BELOW.
This version of the remote server check executed by the
DownloadableFileManager will not check the remote server
for local file: URLs that are presumably on the user's
system. Note that the file-by-file check uses different
code that already did/does not ever check remotely for
such files.

To support developer verification, this version includes two
diagnostic messages that will be shown if --debug-cache is
included on the command line. These two notifications in
the DownloadbleFileManager.java class each include the text:
"This message to be moreved when integrating issue #2060".

Both these messages should be removed before integrating
this branch!