RM 2238: refine ColoringData inheritance hierarchy.
Change getFile() so it does *not* download the file but add
fetchFile() method, which *does* download the file. Use appropriately.

Change which methods are final and abstract to make everything work
correctly whether or not files originate on the server.

Make ColoringData hierarchy thread-safe for retrieving data.

Refactor hierarchy to remove hazardous, unintended, inappropriate
"fall-through" of factory methods to subclasses.

Cosmetic re-ordering of methods.

Update comments.