Addresses #3 - code restructuring
Retooled some of the original code that Bridget put in to make it a bit
more maintainable and so we can easily add new preferences later on
(some code in here like font or axes is hold over from the old
preferences pane but not yet implemented so we can easily pick that up
later). Now also properly saves to the properties file in the cache (as
well as the main window cfg file for that preference).  As a reminder
this includes code from #2 that was required for all of these updates.